feat: Add Synology Photos Importer extensions (#1461)
The Synology Photos application is a useful tool for photo and album
management, designed to operate on the DiskStation Manager (DSM) of
Synology NAS devices. Due to most NAS systems being behind NAT and not
accessible publicly, Synology offers a cloud service that acts as an
intermediary. This service facilitates the transfer of data from various
DTP extensions to the Synology cloud, from which users can subsequently
download their data to their NAS.

This PR includes:
- Synology Photos Importer
- Synology Videos Importer
- Synology Media Importer

---------

Co-authored-by: simonyeh <simonyeh@synology.com>
Co-authored-by: ehung <ehung@synology.com>
Co-authored-by: amychen <amychen@synology.com>