fix: refactor the pagination prefix and token processing (#1476)
Encapsulate pageTokenPrefix and paginationToken within a unified
PageToken class or utility functions to simplify and generalize
pagination logic.
Issue : The logic to validate and extract the pageToken from the request
is not standardized and has repetition for every case which makes it
error prone.

The change was already done for exporting release. 

This change adds generic methods to the StringPaginationToken class and
refactors the export playlist method.